# openapi/v1.yaml
openapi: 3.1.0
info:
  title: registerwatch
  description: Gambling regulators' public registers, one schema per register, with history.
  contact:
    name: registerwatch
    url: https://github.com/mishmish004/RegisterWatch
  license:
    name: All rights reserved
    url: https://choosealicense.com/no-permission/
  version: 0.2.0
servers:
- url: /
  description: The deployment serving this document
paths:
...
  /v1/domains/{domain}:
    get:
      tags:
      - domains
      summary: Where a domain is licensed, and where it is blocked
      description: A leading `www.` is ignored, and a listed subdomain counts as a match of its parent
        (`nj.bet365.com` for `bet365.com`), marked `subdomain`.
      operationId: getDomainStatus
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: domain
        in: path
        required: true
        schema:
          type: string
          maxLength: 253
          description: A hostname; a URL is accepted and reduced to its host
          examples:
          - bet365.com
          title: Domain
        description: A hostname; a URL is accepted and reduced to its host
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Only these jurisdictions; repeat the parameter for several (`?jurisdiction=gb&jurisdiction=de`)
          title: Jurisdiction
        description: Only these jurisdictions; repeat the parameter for several (`?jurisdiction=gb&jurisdiction=de`)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainStatus'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
  /v1/ingest-runs:
    post:
      tags:
      - ingest
      summary: Start an ingest run
      description: 'Answers at once with the run; poll its `url` (also in `Location`). One run at a time
        across every replica: a second is a 409 linking the active one. With `Idempotency-Key`, a repeat
        of the same request within 24 hours returns the original run (200) instead of starting another.'
      operationId: createIngestRun
      security:
      - IngestToken: []
      parameters:
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 255
          - type: 'null'
          description: Any string you choose, unique per intended run
          title: Idempotency-Key
        description: Any string you choose, unique per intended run
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/IngestRunRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: 'A replay of an earlier request with the same `Idempotency-Key`: the original run'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRun'
          headers:
            Cache-Control: *id001
            RateLimit-Policy: *id003
            RateLimit: *id004
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRun'
          headers:
            Cache-Control: *id001
            RateLimit-Policy: *id003
            RateLimit: *id004
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
    get:
      tags:
      - ingest
      summary: List ingest runs, newest first
      operationId: listIngestRuns
      security:
      - IngestToken: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Page size
          default: 100
          title: Limit
        description: Page size
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: '`next_cursor` from the previous page'
          title: Cursor
        description: '`next_cursor` from the previous page'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRunPage'
          headers:
            Cache-Control: *id001
            RateLimit-Policy: *id003
            RateLimit: *id004
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
  /v1/ingest-runs/{id}:
    get:
      tags:
      - ingest
      summary: Get one ingest run
      operationId: getIngestRun
      security:
      - IngestToken: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: The run's id, from its `url`
          title: Id
        description: The run's id, from its `url`
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRun'
          headers:
            Cache-Control: *id001
            RateLimit-Policy: *id003
            RateLimit: *id004
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
...
components:
  schemas:
...
    DomainMatch:
      properties:
        jurisdiction:
          type: string
          title: Jurisdiction
        register:
          type: string
          title: Register
          description: Register slug, also its Postgres schema
        regulator:
          type: string
          title: Regulator
        kind:
          type: string
          enum:
          - licensees
          - blocklist
          title: Kind
        table:
          type: string
          title: Table
        match:
          type: string
          enum:
          - exact
          - subdomain
          title: Match
          description: '`subdomain`: the row lists a subdomain of the domain'
        row:
          $ref: '#/components/schemas/Row'
      type: object
      required:
      - jurisdiction
      - register
      - regulator
      - kind
      - table
      - match
      - row
      title: DomainMatch
      examples:
      - &id006
        jurisdiction: US-NJ
        register: us_nj_dge
        regulator: New Jersey Division of Gaming Enforcement
        kind: licensees
        table: internet_gaming_sites
        match: subdomain
        row:
          id: 9
          first_seen_snapshot_id: 405
          first_seen_at: '2026-10-05T06:01:12Z'
          last_seen_at: '2026-10-07T06:00:41Z'
          values:
            licensee: HARD ROCK HOTEL AND CASINO
            site: nj.bet365.com
            host: nj.bet365.com
            status: authorized
    DomainStatus:
      properties:
        domain:
          type: string
          title: Domain
          description: The hostname checked, lower-cased, without a leading `www.`
        licensed_in:
          items:
            type: string
          type: array
          title: Licensed In
          description: Jurisdictions whose licensee registers list it
        blocked_in:
          items:
            type: string
          type: array
          title: Blocked In
          description: Jurisdictions whose blocklists list it
        matches:
          items:
            $ref: '#/components/schemas/DomainMatch'
          type: array
          title: Matches
      type: object
      required:
      - domain
      - licensed_in
      - blocked_in
      - matches
      title: DomainStatus
      examples:
      - domain: bet365.com
        licensed_in:
        - US-NJ
        blocked_in:
        - CH
        matches:
        - *id006
...
    IngestRun:
      properties:
        id:
          type: string
          title: Id
        status:
          type: string
          enum:
          - queued
          - running
          - succeeded
          - partial
          - failed
          title: Status
          description: '`partial`: some registers incomplete or not started; `failed`: the run itself
            broke'
        requested:
          $ref: '#/components/schemas/IngestRunRequest'
          description: The request as it was received
        registers:
          items:
            type: string
          type: array
          title: Registers
          description: What the request resolved to, in run order
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          description: Why the run failed, e.g. `worker lost`
        results:
          items:
            $ref: '#/components/schemas/IngestRunResult'
          type: array
          title: Results
          description: One per register finished, in run order
        not_started:
          items:
            type: string
          type: array
          title: Not Started
          description: Registers the run stopped before (a shutdown)
        url:
          type: string
          title: Url
      type: object
      required:
      - id
      - status
      - requested
      - registers
      - created_at
      - started_at
      - finished_at
      - error
      - results
      - not_started
      - url
      title: IngestRun
      examples:
      - &id007
        id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
        status: partial
        requested:
          registers: null
          jurisdiction: ch
          force: false
          accept_count_delta: false
        registers:
        - ch_esbk
        - ch_gespa
        created_at: '2026-10-07T06:00:00Z'
        started_at: '2026-10-07T06:00:00Z'
        finished_at: '2026-10-07T06:00:19Z'
        error: null
        results:
        - &id008
          register: ch_esbk
          snapshot_id: 431
          complete: true
          skipped: false
          unchanged: false
          reason: null
          record_count: 1532
          finished_at: '2026-10-07T06:00:12Z'
        not_started:
        - ch_gespa
        url: /v1/ingest-runs/0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
    IngestRunPage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/IngestRun'
          type: array
          title: Data
          description: Newest first
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: IngestRunPage
      examples:
      - data:
        - *id007
        pagination:
          next_cursor: null
          has_more: false
          limit: 100
          total: null
    IngestRunRequest:
      properties:
        registers:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 100
            minItems: 1
          - type: 'null'
          title: Registers
          description: Register slugs; not with `jurisdiction`
        jurisdiction:
          anyOf:
          - type: string
            maxLength: 10
          - type: 'null'
          title: Jurisdiction
          description: Every register of one jurisdiction
        force:
          type: boolean
          title: Force
          description: Fetch even when a good snapshot is recent
          default: false
        accept_count_delta:
          type: boolean
          title: Accept Count Delta
          description: Accept a row count far from the last one
          default: false
      additionalProperties: false
      type: object
      title: IngestRunRequest
      description: 'Which registers to ingest: `registers`, or `jurisdiction`, or neither for all.'
      examples:
      - jurisdiction: ch
      - registers:
        - gb_ukgc
        force: true
      - {}
...
    Pagination:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Pass as `cursor` for the next page; null on the last page
        has_more:
          type: boolean
          title: Has More
        limit:
          type: integer
          title: Limit
          description: Page size this response was read with
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
          description: Matching items; only with `include_total=true` on endpoints that offer it, else
            null
      type: object
      required:
      - next_cursor
      - has_more
      - limit
      title: Pagination
      examples:
      - next_cursor: eyJwIjp7Im8iOjEwMH0sInMiOiI0ZjFkIn0
        has_more: true
        limit: 100
        total: null
...
    Problem:
      additionalProperties: true
      description: RFC 9457 Problem Details. Extension members may be added per type.
      examples:
      - detail: 'limit: Input should be less than or equal to 1000'
        errors:
        - field: limit
          location: query
          message: Input should be less than or equal to 1000
        instance: /v1/registers/gb_ukgc/tables/licences/rows
        request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
        status: 400
        title: Invalid parameter
        type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#invalid-parameter
      properties:
        type:
          description: Stable identifier of the problem type; its docs say what to do
          format: uri
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          description: What went wrong with this request, for a human
          title: Detail
          type: string
        instance:
          description: The request path
          format: uri-reference
          title: Instance
          type: string
        request_id:
          description: Also in the `X-Request-Id` response header and the server's logs
          title: Request Id
          type: string
        errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldError'
            type: array
          - type: 'null'
          default: null
          description: Each invalid parameter, for 400s
          title: Errors
      required:
      - type
      - title
      - status
      - detail
      - instance
      - request_id
      title: Problem
      type: object
  securitySchemes:
    ReadToken:
      type: http
      description: '`READ_TOKEN`. Needed on read routes only when the server sets one; when it does not,
        reads are open. The ingest token reads too.'
      scheme: bearer
    IngestToken:
      type: http
      description: '`INGEST_TOKEN`. Starts and lists ingest runs, and reads. A read token here is a 403.'
      scheme: bearer
  responses:
    BadRequest:
      description: 'A parameter is invalid; `errors` names each one. Types: `invalid-parameter`, `unknown-filter-column`,
        `invalid-cursor`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#invalid-parameter
            title: Invalid parameter
            status: 400
            detail: 'limit: Input should be less than or equal to 1000'
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
            errors:
            - field: limit
              location: query
              message: Input should be less than or equal to 1000
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    Unauthenticated:
      description: 'The bearer token is missing, or is not one this route takes. Types: `unauthenticated`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#unauthenticated
            title: Authentication required
            status: 401
            detail: the bearer token is not valid here
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        WWW-Authenticate:
          description: '`Bearer realm="registerwatch"`, with `error="invalid_token"` when a token was
            sent but is not valid (RFC 6750)'
          schema:
            type: string
        RateLimit-Policy: *id003
        RateLimit: *id004
    Forbidden:
      description: 'The token is valid but cannot do this (a read token on an ingest operation). Types:
        `forbidden`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#forbidden
            title: Not allowed with this token
            status: 403
            detail: a read token cannot start or list ingest runs
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    NotFound:
      description: 'An identifier in the path or body does not exist. Types: `jurisdiction-not-found`,
        `register-not-found`, `table-not-found`, `row-not-found`, `ingest-run-not-found`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#jurisdiction-not-found
            title: Unknown jurisdiction
            status: 404
            detail: 'unknown jurisdiction ''zz''; known: au, be, ca, ...'
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    Conflict:
      description: 'An ingest run is already going; `active_run` links to it. Types: `ingest-in-progress`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#ingest-in-progress
            title: An ingest is already running
            status: 409
            detail: 'an ingest run is already going: /v1/ingest-runs/0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11'
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: &id016
          description: Seconds to wait before retrying
          schema:
            type: integer
        RateLimit-Policy: *id003
        RateLimit: *id004
    UnsupportedMediaType:
      description: 'The body is not `application/json`. Types: `unsupported-media-type`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#unsupported-media-type
            title: Unsupported media type
            status: 415
            detail: send the body as application/json, not text/plain
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    IdempotencyKeyReused:
      description: 'The `Idempotency-Key` was used with a different body. Types: `idempotency-key-reused`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#idempotency-key-reused
            title: Idempotency key reused with a different request
            status: 422
            detail: Idempotency-Key 'cron-2026-10-07' was used with a different body
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    TooManyRequests:
      description: 'Rate limit exceeded; wait `Retry-After` seconds. Types: `rate-limited`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#rate-limited
            title: Too many requests
            status: 429
            detail: 60 requests per minute for /v1/search; retry in 12 s
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    InternalError:
      description: 'An unexpected error; quote `request_id` when reporting it. Types: `internal`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#internal
            title: Internal error
            status: 500
            detail: something went wrong on our side; quote request_id if you report it
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    ServiceUnavailable:
      description: 'The database cannot be reached (retry after `Retry-After` seconds), or ingest is not
        configured on this deployment. Types: `database-unavailable`, `ingest-disabled`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#database-unavailable
            title: Database unavailable
            status: 503
            detail: the database cannot be reached right now; retry shortly
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    QueryTimeout:
      description: 'A query ran past the statement timeout and was cancelled; narrow the request, or retry
        after `Retry-After` seconds. Types: `query-timeout`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#query-timeout
            title: Query took too long
            status: 504
            detail: the query was cancelled for taking too long; narrow it and retry
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    NotModified:
      description: Unchanged since the `ETag` sent in `If-None-Match`; no body
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
        RateLimit-Policy: *id003
        RateLimit: *id004
  headers:
    ETag:
      description: This representation's version (weak). Send it back in `If-None-Match` and the answer
        is a 304 with no body while the data is unchanged. New data shows within 30 seconds of its snapshot.
      schema:
        type: string
      example: W/"9f2c5d0e..."
    Cache-Control:
      description: '`public, max-age=300` when reads are open, `private, max-age=300` when the deployment
        requires a read token; `no-store` on ingest operations'
      schema:
        type: string
...